requirements.txt
streamlit
pydantic
google-generativeai
python-dotenv